RHINENG-15783: remove baselines api
