+++ base/deprecations/deprecations.go
@@ -2,7 +2,6 @@ package deprecations
 
 import (
 	"app/base/utils"
-	"strings"
 	"time"
 
 	"github.com/gin-gonic/gin"
@@ -19,18 +18,6 @@ func DeprecateLimit() Deprecation {
 				return true
 			}
 			return false
-		},
-	}
-}
-
-// Deprecate baselines api
-func DeprecateBaselines() Deprecation {
-	return apiDeprecation{
-		deprecationTimestamp: time.Date(2025, 3, 1, 0, 0, 0, 0, time.UTC),
-		message:              "baselines are deprecated - https://access.redhat.com/articles/7097146",
-		shouldDeprecate: func(c *gin.Context) bool {
-			handlerName := c.HandlerName()
-			return strings.Contains(strings.ToLower(handlerName), "baseline")
 		},
 	}
 }
