+++ docs/docs_test.go
@@ -19,7 +19,6 @@ func TestValidateOpenAPI3DocStr(t *testing.T) {
 
 func TestFilterOpenAPIPaths1(t *testing.T) {
 	nRemovedPaths := filterOpenAPI(EndpointsConfig{
-		EnableBaselines: true,
 		EnableTemplates: true,
 	}, openAPIPath, "/tmp/openapi-filter-test.json")
 	assert.Equal(t, 0, nRemovedPaths)
@@ -27,8 +26,7 @@ func TestFilterOpenAPIPaths1(t *testing.T) {
 
 func TestFilterOpenAPIPaths2(t *testing.T) {
 	nRemovedPaths := filterOpenAPI(EndpointsConfig{
-		EnableBaselines: false,
 		EnableTemplates: false,
 	}, openAPIPath, "/tmp/openapi-filter-test.json")
-	assert.Equal(t, 12, nRemovedPaths)
+	assert.Equal(t, 6, nRemovedPaths)
 }
