--- manager/controllers/baseline_delete_test.go
@@ -1,46 +1,0 @@
-package controllers
-
-import (
-	"app/base/core"
-	"app/base/database"
-	"app/base/utils"
-	"fmt"
-	"net/http"
-	"testing"
-
-	"github.com/stretchr/testify/assert"
-)
-
-func TestBaselineDelete(t *testing.T) {
-	core.SetupTest(t)
-	var inventoryIDs = []string{
-		"00000000-0000-0000-0000-000000000005",
-		"00000000-0000-0000-0000-000000000006",
-		"00000000-0000-0000-0000-000000000007"}
-	baselineID := database.CreateBaseline(t, "", inventoryIDs, nil)
-
-	w := CreateRequestRouterWithPath("GET", "/:baseline_id", fmt.Sprintf(`%v`, baselineID), "", nil, "",
-		BaselineDeleteHandler)
-
-	var resp DeleteBaselineResponse
-	CheckResponse(t, w, http.StatusOK, &resp)
-	database.CheckBaselineDeleted(t, resp.BaselineID)
-}
-
-func TestBaselineDeleteNonExisting(t *testing.T) {
-	core.SetupTest(t)
-	w := CreateRequestRouterWithPath("GET", "/:baseline_id", "88888", "", nil, "", BaselineDeleteHandler)
-
-	var errResp utils.ErrorResponse
-	CheckResponse(t, w, http.StatusNotFound, &errResp)
-	assert.Equal(t, "baseline not found", errResp.Error)
-}
-
-func TestBaselineDeleteInvalid(t *testing.T) {
-	core.SetupTestEnvironment()
-	w := CreateRequestRouterWithPath("GET", "/:baseline_id", "invalidBaseline", "", nil, "", BaselineDeleteHandler)
-
-	var errResp utils.ErrorResponse
-	CheckResponse(t, w, http.StatusBadRequest, &errResp)
-	assert.Equal(t, "Invalid baseline_id: invalidBaseline", errResp.Error)
-}
