--- manager/controllers/baseline_systems_export.go
@@ -1,56 +1,0 @@
-package controllers
-
-import (
-	"app/base/utils"
-
-	"github.com/gin-gonic/gin"
-)
-
-// nolint: lll
-// @Summary Export systems belonging to a baseline
-// @Description  Export systems applicable to a baseline. Export endpoints are not paginated.
-// @ID exportBaselineSystems
-// @Security RhIdentity
-// @Accept   json
-// @Produce  json,text/csv
-// @Param    baseline_id    path    int     true    "Baseline ID"
-// @Param    search         query   string  false   "Find matching text"
-// @Param    filter[display_name]           query   string  false "Filter"
-// @Param    filter[os]           			query   string  false "Filter"
-// @Param    tags           query   []string  false "Tag filter"
-// @Param    filter[group_name] 									query []string 	false "Filter systems by inventory groups"
-// @Param    filter[system_profile][sap_system]						query bool  	false "Filter only SAP systems"
-// @Param    filter[system_profile][sap_sids]						query []string  false "Filter systems by their SAP SIDs"
-// @Param    filter[system_profile][ansible]						query string 	false "Filter systems by ansible"
-// @Param    filter[system_profile][ansible][controller_version]	query string 	false "Filter systems by ansible version"
-// @Param    filter[system_profile][mssql]							query string 	false "Filter systems by mssql version"
-// @Param    filter[system_profile][mssql][version]					query string 	false "Filter systems by mssql version"
-// @Success 200 {array} BaselineSystemsDBLookup
-// @Failure 400 {object} utils.ErrorResponse
-// @Failure 404 {object} utils.ErrorResponse
-// @Failure 415 {object} utils.ErrorResponse
-// @Failure 500 {object} utils.ErrorResponse
-// @Router /export/baselines/{baseline_id}/systems [get]
-func BaselineSystemsExportHandler(c *gin.Context) {
-	account := c.GetInt(utils.KeyAccount)
-	groups := c.GetStringMapString(utils.KeyInventoryGroups)
-
-	query, err := queryBaselineSystems(c, account, groups)
-	if err != nil {
-		return
-	} // Error handled in method itself
-
-	query, err = ExportListCommon(query, c, BaselineSystemOpts)
-	if err != nil {
-		return
-	} // Error handled in method itself
-
-	var baselineSystems []BaselineSystemsDBLookup
-	err = query.Find(&baselineSystems).Error
-	if err != nil {
-		LogAndRespError(c, err, err.Error())
-		return
-	}
-
-	OutputExportData(c, baselineSystems)
-}
